Fixed bug with BACKEND_URL; Fixed React error in front end
